pyproject.toml
[project]
name = "simple_llm_api"
version = "0.1.0"
description = "A simple python chat completion api for finetuned small language model (gguf) file"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "huggingface-hub>=0.34.4",
    "llama-cpp-python>=0.3.16",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
]
